Current progress putting in the new connection implementation.
