+++ ReadUs/RedisSingleInstanceConnectionPool.cs
@@ -34,7 +34,7 @@ public class RedisSingleInstanceConnectionPool : RedisConnectionPool
             return connection;
         }
 
-        var newConnection = new RedisConnection(_configuration);
+        var newConnection = new OldRedisConnection(_configuration);
 
         _allConnections.Add(newConnection);
